requirements.txt
requests>=2.31.0
PySide6>=6.6.0
PyPDF2>=3.0.0
python-docx>=1.1.0
numpy>=1.26.0
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
pydicom>=2.4.4
Pillow>=10.2.0
nibabel>=5.2.1
sounddevice>=0.4.6
soundfile>=0.12.1
pyttsx3>=2.90
pyinstaller>=6.2.0
